Modificaciones de type en Asignatura

+++ src/main/java/classroom/Asignatura.java
@@ -14,24 +14,22 @@ public class Asignatura {
 
     //Asignatura(int codigoInterno,Tipo tipo) { this("Sin nombre", codigoInterno, 0,tipo);  }
 
-    Asignatura(int codigoExterno,Tipo tipo) {
-
+    Asignatura(long codigoExterno,Tipo tipo) {
         this("Sin nombre", 0, codigoExterno,tipo);
     }
 
     Asignatura(String nombre, Tipo tipo) {
-
         this(nombre, 0, 0,tipo);
     }
 
-    Asignatura(String nombre, int codigoInterno, int codigoExterno,Tipo tipo) {
+    Asignatura(String nombre, int codigoInterno, long codigoExterno,Tipo tipo) {
         this.nombre = nombre;
         this.codigoInterno = codigoInterno;
         this.codigoExterno = codigoExterno;
         this.tipo = tipo;
     }
 
-    void cambiarDatos(int codigoInterno, int codigoExterno, String nombre,Tipo tipo) {
+    void cambiarDatos(int codigoInterno, long codigoExterno, String nombre,Tipo tipo) {
         this.codigoInterno = codigoInterno;
         this.codigoExterno = codigoExterno;
         this.nombre = nombre;
@@ -51,7 +49,7 @@ public class Asignatura {
     public void setCodigoInterno(int codigoInterno) {
         this.codigoInterno = codigoInterno;
     }
-    
+
     public void setCodigoInterno(double codigoInterno) {
         this.codigoInterno = (int) codigoInterno;
     }
